+++ public/Localizations/en.json
@@ -28,7 +28,7 @@
     "give_access_to_new_api": "Give access to new API",
     "new_org_accordion_content_text": "Select which users will have access by clicking the plus sign. You can also add new.",
     "businesses_previously_delegated_to": "Businesses previously delegated to:",
-    "businesses_going_to_get_access":"Businesses that will receive the right",
+    "businesses_going_to_get_access":"Businesses that will receive the right:",
     "previous":"Previous",
     "next": "Next",
     "add_new_business": "Add new business"
