+++ public/Localizations/no_nn.json
@@ -27,7 +27,7 @@
     "give_access_to_new_api": "Gje tilgang til nytt API",
     "new_org_accordion_content_text": "Vel kva brukarar som skal få tilgang ved å klikka pluss-teiknet. Du kan også leggja til ny.",
     "businesses_previously_delegated_to": "Verksemder det tidlegare er delegert til:",
-    "businesses_going_to_get_access":"Verksemder som skal få rett",
+    "businesses_going_to_get_access":"Verksemder som skal få rett:",
     "previous":"Forrige",
     "next": "Neste",
     "add_new_business": "Add new business"
