+++ src/components/Common/NewDelegationAccordion/index.ts
@@ -1,0 +1,1 @@
+export { NewDelegationAccordion, AccordionButtonType } from './NewDelegationAccordion';
