--- src/components/NewApiDelegationPage/NewApiDelegationAccordion/index.ts
@@ -1,1 +1,0 @@
-export { NewApiDelegationAccordion, AccordionButtonType } from './NewApiDelegationAccordion';
