+++ src/components/NewApiDelegationPage/NewApiDelegationPage.module.css
@@ -39,7 +39,7 @@
 
 
 .searchField{
-  width: 100%;
+  width: 47%;
 }
 
 .navButtonContainer {
