--- src/components/NewOrgDelegationPage/NewApiDelegationAccordion/NewOrgDelegationAccordion.module.css
@@ -1,3 +1,0 @@
-.newApiAccordionContent{
-  margin: 2rem;
-}
