--- src/components/NewOrgDelegationPage/NewApiDelegationAccordion/index.ts
@@ -1,1 +1,0 @@
-export { NewOrgDelegationAccordion, AccordionButtonType } from './NewOrgDelegationAccordion';
