+++ src/components/NewOrgDelegationPage/NewOrgDelegationPage.module.css
@@ -54,6 +54,6 @@
 }
 
 .accordionScrollContainer {
-  max-height: 300px;
+  height: 300px;
   overflow-y: auto;
 }
